finish tournament starting implementation
