+++ src/lib/constants.ts
@@ -11,6 +11,9 @@ export const TEAM_NAME_MIN_LENGTH = 3 as const;
 export const TEAM_NAME_MAX_LENGTH = 25 as const;
 
 export const SUBMISSION_MAX_FILE_SIZE = 10485760 as const; // 10 MB (10 * 1024 * 1024 bytes)
+export const TABLE_SEATS_MAX_COUNT = 8;
+export const TABLE_SEATS_MIN_COUNT = 2;
+export const TABLE_SEATS_MIN_START = 5;
 
 export const UCSB_POKER_TOURNEY_ID =
   "f6fd507b-42fb-4fba-a0d3-e9ded05aeca5" as const; // hard code for now, implement different tournaments later
