--- todo.md
@@ -1,1 +1,0 @@
-FIX USERS RLS POLICY TO NOT ALLOW CHANGES TO is_admin
